Noticed that I made a boo-boo
Must mention that a .json file is required when pushing this to a remote device to handle files. Will add section later to allow for the download and upload of .json files for the fun of it. It will still technically be useless for first runs, as the .json file is required to make the request. It will enable the user to download and store other .json files in a separate folder if needed.